stock-market-app/requirements.txt
flask
requests
python-dotenv
plotly